+++ embed_video/settings.py
@@ -1,6 +1,5 @@
 from django.conf import settings
 
-
 EMBED_VIDEO_BACKENDS = getattr(
     settings,
     "EMBED_VIDEO_BACKENDS",