+++ embed_video/tests/__init__.py
@@ -1,4 +1,5 @@
 import os
+
 import django
 
 os.environ["DJANGO_SETTINGS_MODULE"] = "embed_video.tests.django_settings"