+++ embed_video/tests/templatetags/tests_embed_video_tags.py
@@ -3,8 +3,8 @@ import urllib.parse as urlparse
 from unittest import TestCase
 from unittest.mock import Mock, patch
 
-from django.template import TemplateSyntaxError
 from django.http import HttpRequest
+from django.template import TemplateSyntaxError
 from django.template.base import Template
 from django.template.context import RequestContext
 from django.test.client import RequestFactory
@@ -59,7 +59,7 @@ class EmbedTestCase(TestCase):
             template,
             '<iframe width="960" height="720" '
             'src="https://www.youtube.com/embed/jsrRJyHBvzw?wmode=opaque" '
-            'frameborder="0" allowfullscreen></iframe>',
+            'loading="lazy" frameborder="0" allowfullscreen></iframe>',
         )
 
     def test_embed_invalid_url(self):
@@ -95,7 +95,7 @@ class EmbedTestCase(TestCase):
             template,
             '<iframe width="960" height="720" '
             'src="https://www.youtube.com/embed/jsrRJyHBvzw?wmode=opaque" '
-            'frameborder="0" allowfullscreen></iframe>',
+            'loading="lazy" frameborder="0" allowfullscreen></iframe>',
         )
 
     def test_direct_embed_tag_with_default_size(self):
@@ -107,7 +107,7 @@ class EmbedTestCase(TestCase):
             template,
             '<iframe width="480" height="360" '
             'src="https://www.youtube.com/embed/jsrRJyHBvzw?wmode=opaque" '
-            'frameborder="0" allowfullscreen></iframe>',
+            'loading="lazy" frameborder="0" allowfullscreen></iframe>',
         )
 
     def test_direct_embed_invalid_url(self):
@@ -128,7 +128,7 @@ class EmbedTestCase(TestCase):
             template,
             '<iframe width="800" height="800" '
             'src="https://www.youtube.com/embed/jsrRJyHBvzw?wmode=opaque" '
-            'frameborder="0" allowfullscreen></iframe>',
+            'loading="lazy" frameborder="0" allowfullscreen></iframe>',
         )
 
     def test_wrong_size(self):
@@ -207,7 +207,7 @@ class EmbedTestCase(TestCase):
             template,
             '<iframe width="80%" height="30%" '
             'src="https://player.vimeo.com/video/72304002" '
-            'frameborder="0" allowfullscreen></iframe>',
+            'loading="lazy" frameborder="0" allowfullscreen></iframe>',
         )
 
     def test_allow_spaces_in_size(self):
@@ -219,7 +219,7 @@ class EmbedTestCase(TestCase):
             template,
             '<iframe width="80%" height="300" '
             'src="https://player.vimeo.com/video/72304002" '
-            'frameborder="0" allowfullscreen></iframe>',
+            'loading="lazy" frameborder="0" allowfullscreen></iframe>',
         )
 
     def test_embed_with_query(self):
@@ -259,7 +259,7 @@ class EmbedTestCase(TestCase):
             output_without_url,
             '<iframe width="480" height="360" '
             'src="URL" '
-            'frameborder="0" allowfullscreen></iframe>',
+            'loading="lazy" frameborder="0" allowfullscreen></iframe>',
         )
 
     def test_set_options(self):
@@ -271,7 +271,7 @@ class EmbedTestCase(TestCase):
             template,
             '<iframe width="300" height="200" '
             'src="https://www.youtube.com/embed/jsrRJyHBvzw?rel=1" '
-            'frameborder="0" allowfullscreen></iframe>',
+            'loading="lazy" frameborder="0" allowfullscreen></iframe>',
         )
 
     def test_size_as_variable(self):
@@ -285,7 +285,7 @@ class EmbedTestCase(TestCase):
             template,
             '<iframe width="500" height="200" '
             'src="https://www.youtube.com/embed/jsrRJyHBvzw?wmode=opaque" '
-            'frameborder="0" allowfullscreen></iframe>',
+            'loading="lazy" frameborder="0" allowfullscreen></iframe>',
         )